separates key fetch and spotify auth into own services
