+++ app/src/main/java/com/niehusst/partyq/services/SpotifyAuthenticationService.kt
@@ -1,21 +1,39 @@
 package com.niehusst.partyq.services
 
-import android.util.Base64
+import android.content.Context
+import com.spotify.android.appremote.api.ConnectionParams
+import com.spotify.android.appremote.api.Connector
+import com.spotify.android.appremote.api.SpotifyAppRemote
+import timber.log.Timber
 
 class SpotifyAuthenticationService {
-    // TODO: make another service just for handling keys?? dont forget to change c
 
-    companion object {
+    private val clientId = KeyFetchService.getSpotifyKey()
+    private val redirectUri = "com.niehusst.partyq://callback"
+    var spotifyAppRemote: SpotifyAppRemote? = null
 
-        init {
-            System.loadLibrary("keys")
-        }
+    private fun authenticateWithSpotfiy(
+        context: Context?,
+        onConnectCallback: (() -> Unit)?,
+        onFailCallback: (() -> Unit)?
+    ) {
+        val connectionParams = ConnectionParams.Builder(clientId)
+            .setRedirectUri(redirectUri)
+            .showAuthView(true)
+            .build()
 
-        private external fun getSpotifyClientKey(): String
+        SpotifyAppRemote.connect(context, connectionParams, object : Connector.ConnectionListener {
+            override fun onConnected(appRemote: SpotifyAppRemote) {
+                spotifyAppRemote = appRemote
+                Timber.d("Connected to Spotify!")
+                onConnectCallback?.invoke()
+            }
 
-        fun getSpotifyKey(): String {
-            // TODO: return env var if in circleci?
-            return String(Base64.decode(getSpotifyClientKey(), Base64.DEFAULT))
-        }
+            override fun onFailure(throwable: Throwable) {
+                Timber.e("Failed to connect to Spotify:\n $throwable")
+                // Something went wrong when attempting to connect
+                onFailCallback?.invoke()
+            }
+        })
     }
 }
