+++ app/src/main/java/com/niehusst/partyq/ui/spotifyLogin/SpotifyLoginFragment.kt
@@ -7,14 +7,12 @@ import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.findNavController
-import androidx.navigation.ui.NavigationUI
 import com.niehusst.partyq.R
 import com.niehusst.partyq.databinding.SpotifyLoginFragmentBinding
-import com.niehusst.partyq.services.SpotifyAuthenticationService
+import com.niehusst.partyq.services.KeyFetchService
 import com.spotify.android.appremote.api.ConnectionParams
 import com.spotify.android.appremote.api.Connector
 import com.spotify.android.appremote.api.SpotifyAppRemote
-import kotlinx.android.synthetic.main.activity_main.*
 import timber.log.Timber
 
 class SpotifyLoginFragment : Fragment() {
@@ -35,7 +33,7 @@ class SpotifyLoginFragment : Fragment() {
 
         binding.spotifyAuthButton.setOnClickListener {
             // TODO: auth with spotify
-            connectToSpotfiy()
+            connectToSpotify()
         }
 
         binding.infoButton.setOnClickListener {
@@ -44,29 +42,7 @@ class SpotifyLoginFragment : Fragment() {
     }
 
     // TODO: this should go in its own service, which should be accessed in the viewModel
-    private val clientId = SpotifyAuthenticationService.getSpotifyKey()
-    private val redirectUri = "com.niehusst.partyq://callback"
-    private var spotifyAppRemote: SpotifyAppRemote? = null
-    private fun connectToSpotfiy() {
-        val connectionParams = ConnectionParams.Builder(clientId)
-            .setRedirectUri(redirectUri)
-            .showAuthView(true)
-            .build()
+    private fun connectToSpotify() {
 
-        SpotifyAppRemote.connect(requireContext(), connectionParams, object : Connector.ConnectionListener {
-            override fun onConnected(appRemote: SpotifyAppRemote) {
-                spotifyAppRemote = appRemote
-                Timber.d("Connected to Spotify!")
-                // Now you can start interacting with App Remote
-                // TODO: save connection info for PartyActivity
-                Toast.makeText(requireContext(), "Connect success", Toast.LENGTH_SHORT).show()
-            }
-
-            override fun onFailure(throwable: Throwable) {
-                Timber.e(throwable)
-                // Something went wrong when attempting to connect
-                Toast.makeText(requireContext(), "Error connecting to Spotify app", Toast.LENGTH_LONG).show()
-            }
-        })
     }
 }
